+++ plugins/frontend/backstage-plugin-iframe/src/components/types.ts
@@ -11,6 +11,23 @@ export type IFrameProps = {
   class?: string;
 };
 
+export type IFrameComponentProps = {
+  src?: string;
+  srcFromAnnotation?: string;
+  title?: string;
+  height?: string;
+  width?: string;
+  class?: string;
+};
+
+export type IFrameFromAnnotationProps = {
+  srcFromAnnotation: string;
+  title?: string;
+  height?: string;
+  width?: string;
+  class?: string;
+};
+
 export type IFrameContentProps = {
   iframe: IFrameProps;
   title?: string;