+++ plugins/scaffolder-actions/scaffolder-backend-module-http-request/src/actions/run/backstageRequest.ts
@@ -15,7 +15,11 @@
  */
 
 import { createTemplateAction } from '@backstage/plugin-scaffolder-backend';
-import { generateBackstageUrl, http } from './helpers';
+import {
+  generateBackstageUrl,
+  http,
+  getObjFieldCaseInsensitively,
+} from './helpers';
 import { HttpOptions, Headers, Params, Methods, Body } from './types';
 import { Config } from '@backstage/config';
 
@@ -127,7 +131,12 @@ export function createHttpBackstageAction(options: { config: Config }) {
         body: inputBody,
       };
 
-      if (token) {
+      const authToken = getObjFieldCaseInsensitively(
+        input.headers,
+        'authorization',
+      );
+
+      if (token && !authToken) {
         ctx.logger.info(`Token is defined. Setting authorization header.`);
         httpOptions.headers.authorization = `Bearer ${token}`;
       }