+++ plugins/backend/catalog-backend-module-okta/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @roadiehq/catalog-backend-module-okta
 
+## 0.2.4
+
+### Patch Changes
+
+- 8e003da: Fix readme
+
+## 0.2.3
+
+### Patch Changes
+
+- 151b46b: bump to latest backstage package versions
+
+## 0.2.2
+
+### Patch Changes
+
+- 3c7303c: Allow using naming strategy for group -> user association.
+
+## 0.2.1
+
+### Patch Changes
+
+- 9539caf: Release plugin to trigger rebuild of package.
+
 ## 0.2.0
 
 ### Minor Changes