+++ plugins/frontend/backstage-plugin-github-pull-requests/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @roadiehq/backstage-plugin-github-pull-requests
 
+## 2.2.7
+
+### Patch Changes
+
+- 99153fe: Move react-router and react-router-dom dependencies to peerDependencies because of the migration to the stabel version of react-router in backstage/backstage. See the migration guide [here](https://backstage.io/docs/tutorials/react-router-stable-migration#for-plugin-authors)
+
+## 2.2.6
+
+### Patch Changes
+
+- 151b46b: bump to latest backstage package versions
+
+## 2.2.5
+
+### Patch Changes
+
+- dedfbc5: Fix label styling in `PullRequestsStatsCard`
+
+## 2.2.4
+
+### Patch Changes
+
+- c9cfaad: Release all plugins after fixing typescript exports issue.
+
 ## 2.2.3
 
 ### Patch Changes