+++ plugins/frontend/backstage-plugin-prometheus/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @roadiehq/backstage-plugin-prometheus
 
+## 2.1.1
+
+### Patch Changes
+
+- 151b46b: bump to latest backstage package versions
+
+## 2.1.0
+
+### Minor Changes
+
+- 5b305b9: Add ability to use multiple instances over proxy annotation.
+
+## 2.0.3
+
+### Patch Changes
+
+- c9cfaad: Release all plugins after fixing typescript exports issue.
+
 ## 2.0.2
 
 ### Patch Changes