+++ plugins/frontend/backstage-plugin-shortcut/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @roadiehq/backstage-plugin-shortcut
 
+## 1.0.11
+
+### Patch Changes
+
+- 151b46b: bump to latest backstage package versions
+
+## 1.0.10
+
+### Patch Changes
+
+- c9cfaad: Release all plugins after fixing typescript exports issue.
+
 ## 1.0.9
 
 ### Patch Changes