+++ plugins/home/backstage-plugin-home-markdown/CHANGELOG.md
@@ -1,5 +1,17 @@
 ## 1.1.0
 
+## 2.0.6
+
+### Patch Changes
+
+- 151b46b: bump to latest backstage package versions
+
+## 2.0.5
+
+### Patch Changes
+
+- c9cfaad: Release all plugins after fixing typescript exports issue.
+
 ## 2.0.4
 
 ### Patch Changes