+++ plugins/home/backstage-plugin-home-rss/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @roadiehq/backstage-plugin-home-rss
 
+## 1.0.6
+
+### Patch Changes
+
+- 151b46b: bump to latest backstage package versions
+
+## 1.0.5
+
+### Patch Changes
+
+- c9cfaad: Release all plugins after fixing typescript exports issue.
+
 ## 1.0.4
 
 ### Patch Changes